remove unused plugins
