+++ src/packages/core/src/index.ts
@@ -9,7 +9,6 @@ export * from './hook-manager';
 export * from './metadata';
 export * from './metadata-service';
 export * from './open-telemetry';
-export * from './request-context';
 export * from './schema-builder';
 export * from './types';
 export * from './utils/create-or-update-entities';
