+++ src/packages/server/src/apollo-plugins/index.ts
@@ -1,7 +1,5 @@
-export { ClearDataLoaderCache } from './clear-data-loader-cache';
 export { LogErrors } from './log-errors';
 export { LogRequests } from './log-requests';
-export { MutexRequestsInDevelopment } from './mutex-requests-in-development';
 export { corsPlugin } from './cors';
 export { dedupeGraphQL } from './graphql-deduplicator';
